refactor(frontend): update contentType dialogs
